requirements.txt
python-amazon-sp-api
python-dotenv
psycopg2-binary
httpx
